Shrink resolver `mod.rs` a little
The resolver's `mod.rs` has become very complex, with `ResolverState` being around 3k lines. Move some of this code out of there to reduce complexity.
